+++ src-ui/app/app.module.ts
@@ -243,6 +243,7 @@ import { FrameLimiterSelectorComponent } from './components/frame-limiter-select
 import { FrameLimiterService } from './services/frame-limiter.service';
 import { FrameLimitAutomationsService } from './services/frame-limit-automations.service';
 import { FrameLimiterAddApplicationModalComponent } from './views/dashboard-view/views/frame-limiter-view/modals/frame-limiter-add-application-modal/frame-limiter-add-application-modal.component';
+import { OscAddressAutocompleteComponent } from './components/osc-script-simple-editor/osc-address-autocomplete/osc-address-autocomplete.component';
 
 [
   localeEN,
@@ -394,6 +395,7 @@ export function createTranslateLoader(http: HttpClient) {
     FrameLimiterViewComponent,
     FrameLimiterSelectorComponent,
     FrameLimiterAddApplicationModalComponent,
+    OscAddressAutocompleteComponent,
   ],
   exports: [SelectBoxComponent],
   imports: [